requirements.txt
Flask==2.3.3
kubernetes==27.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
requests==2.31.0
urllib3==2.0.7
python-dateutil==2.8.2
six==1.16.0
PyYAML==6.0.1
google-auth==2.23.4
cachetools==5.3.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
rsa==4.9
oauthlib==3.2.2
requests-oauthlib==1.3.1
websocket-client==1.6.4
pytz==2023.3